+++ waspc/e2e-tests/Tests/WaspNewSnapshotTest.hs
@@ -1,0 +1,10 @@
+module Tests.WaspNewSnapshotTest (waspNewSnapshotTest) where
+
+import SnapshotTest (SnapshotTest, makeSnapshotTest)
+import SnapshotTest.ShellCommands (createSnapshotWaspProjectFromMinimalStarter)
+
+waspNewSnapshotTest :: SnapshotTest
+waspNewSnapshotTest =
+  makeSnapshotTest
+    "wasp-new"
+    [createSnapshotWaspProjectFromMinimalStarter]