+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/sdk/wasp/auth/password.ts
@@ -1,0 +1,1 @@
+export { hashPassword, verifyPassword } from "@wasp.sh/lib-auth/sdk";