+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/sdk/wasp/src/features/auth/pages/ManualSignupPage.tsx
@@ -1,0 +1,91 @@
+import { useState } from "react";
+import {
+  FormError,
+  FormInput,
+  FormItemGroup,
+  FormLabel,
+  signup,
+} from "wasp/client/auth";
+// Missing SubmitButton export
+// import { SubmitButton } from 'wasp/client/auth'
+import { useForm } from "react-hook-form";
+import { Alert } from "../../../components/Alert";
+import { Button } from "../../../components/Button";
+import { FeatureContainer } from "../../../components/FeatureContainer";
+
+export const ManualSignupPage = () => {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<{
+    email: string;
+    password: string;
+    address: string;
+  }>();
+  const [message, setMessage] = useState<{
+    type: "success" | "error";
+    text: string;
+  } | null>(null);
+
+  const onSubmit = handleSubmit(async (data) => {
+    try {
+      const result = await signup(data);
+      console.error("result", result);
+      if (result.success) {
+        setMessage({
+          type: "success",
+          text: "You've signed up successfully! Check your email for the confirmation link.",
+        });
+      } else {
+        setMessage({
+          type: "error",
+          text: "Signup failed. Please try again.",
+        });
+      }
+    } catch (error: any) {
+      const { message, data } = error.data;
+      setMessage({
+        type: "error",
+        text: `${message}: ${data.message}`,
+      });
+    }
+  });
+
+  return (
+    <FeatureContainer>
+      <div className="space-y-4">
+        <h2 className="feature-title">Manual Signup Form</h2>
+        <form
+          onSubmit={onSubmit}
+          className="card"
+          data-testid="manual-signup-form"
+        >
+          {message && (
+            <Alert variant={message.type} className="mb-4">
+              <span data-testid="message">{message.text}</span>
+            </Alert>
+          )}
+          <FormItemGroup>
+            <FormLabel>E-mail</FormLabel>
+            <FormInput type="email" {...register("email")} />
+            <FormError>{errors.email?.message}</FormError>
+          </FormItemGroup>
+          <FormItemGroup>
+            <FormLabel>Password</FormLabel>
+            <FormInput type="password" {...register("password")} />
+            <FormError>{errors.password?.message}</FormError>
+          </FormItemGroup>
+          <FormItemGroup>
+            <FormLabel>Address</FormLabel>
+            <FormInput {...register("address")} />
+            <FormError>{errors.address?.message}</FormError>
+          </FormItemGroup>
+          <FormItemGroup>
+            <Button type="submit">Sign up</Button>
+          </FormItemGroup>
+        </form>
+      </div>
+    </FeatureContainer>
+  );
+};