+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/actions/requestUppercaseText.ts
@@ -1,0 +1,13 @@
+import { prisma } from 'wasp/server'
+
+import { requestUppercaseText } from '../../../../../src/features/jobs/uppercaseText'
+
+
+export default async function (args, context) {
+  return (requestUppercaseText as any)(args, {
+    ...context,
+    entities: {
+      UppercaseTextRequest: prisma.uppercaseTextRequest,
+    },
+  })
+}