+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/actions/voidToStringNoAuth.ts
@@ -1,0 +1,13 @@
+import { prisma } from 'wasp/server'
+
+import { voidToStringNoAuth } from '../../../../../src/rpcTests/operations/definitions'
+
+
+export default async function (args, context) {
+  return (voidToStringNoAuth as any)(args, {
+    ...context,
+    entities: {
+      Task: prisma.task,
+    },
+  })
+}