+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/jobs/core/allJobs.ts
@@ -1,0 +1,7 @@
+
+// This module imports all jobs and is imported by the server to ensure
+// any schedules that are not referenced are still loaded by NodeJS.
+
+import '../uppercaseTextJob.js'
+import '../mySpecialJob.js'
+import '../mySpecialScheduledJob.js'