+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/routes/operations/getAnythingNoAuth.js
@@ -1,0 +1,4 @@
+import { createQuery } from '../../middleware/operations.js'
+import getAnythingNoAuth from '../../queries/getAnythingNoAuth.js'
+
+export default createQuery(getAnythingNoAuth)