+++ waspc/src/Wasp/Generator/ServerGenerator.hs
@@ -6,7 +6,7 @@
 module Wasp.Generator.ServerGenerator
   ( genServer,
     operationsRouteInRootRouter,
-    npmDepsForWasp,
+    npmDepsFromWasp,
   )
 where
 
@@ -47,7 +47,9 @@ import Wasp.Generator.DepVersions (superjsonVersion, typescriptVersion)
 import Wasp.Generator.FileDraft (FileDraft, createTextFileDraft)
 import Wasp.Generator.Monad (Generator)
 import qualified Wasp.Generator.Monad as Generator
+import Wasp.Generator.NpmDependencies (NpmDepsForPackage (peerDependencies))
 import qualified Wasp.Generator.NpmDependencies as N
+import Wasp.Generator.NpmWorkspaces (serverPackageName)
 import Wasp.Generator.ServerGenerator.ApiRoutesG (genApis)
 import Wasp.Generator.ServerGenerator.AuthG (genAuth)
 import qualified Wasp.Generator.ServerGenerator.Common as C
@@ -69,16 +71,16 @@ import Wasp.Util ((<++>))
 
 genServer :: AppSpec -> Generator [FileDraft]
 genServer spec = do
-  npmDeps <- npmDepsForWasp spec <$> Generator.getWaspLibs
+  npmDeps <- npmDepsFromWasp spec <$> Generator.getWaspLibs
   sequence
     [ genFileCopy [relfile|README.md|],
       genRollupConfigJs spec,
       genTsConfigJson spec,
       genPackageJson spec npmDeps,
-      genNpmrc,
       genGitignore,
       genNodemon
     ]
+    <++> genNpmrc spec
     <++> genSrcDir spec
     <++> genDotEnv spec
     <++> genJobs spec
@@ -124,7 +126,7 @@ genTsConfigJson spec = do
     srcTsConfigPath :: Path' (Rel C.ServerRootDir) (File SrcTsConfigFile) =
       waspProjectDirFromAppComponentDir </> AS.srcTsConfigPath spec
 
-genPackageJson :: AppSpec -> N.NpmDepsForWasp -> Generator FileDraft
+genPackageJson :: AppSpec -> N.NpmDepsFromWasp -> Generator FileDraft
 genPackageJson spec waspDependencies = do
   combinedDependencies <- N.genNpmDepsForPackage spec waspDependencies
   return $
@@ -133,7 +135,8 @@ genPackageJson spec waspDependencies = do
       (C.asServerFile [relfile|package.json|])
       ( Just $
           object
-            [ "depsChunk" .= N.getDependenciesPackageJsonEntry combinedDependencies,
+            [ "packageName" .= serverPackageName spec,
+              "depsChunk" .= N.getDependenciesPackageJsonEntry combinedDependencies,
               "devDepsChunk" .= N.getDevDependenciesPackageJsonEntry combinedDependencies,
               "nodeVersionRange" .= (">=" <> show NodeVersion.oldestWaspSupportedNodeVersion),
               "startProductionScript"
@@ -151,49 +154,62 @@ getPackageJsonPrismaField spec = object $ [] <> seedEntry
   where
     seedEntry = maybeToList $ Just . ("seed" .=) =<< getPackageJsonPrismaSeedField spec
 
-npmDepsForWasp :: AppSpec -> [WaspLib.WaspLib] -> N.NpmDepsForWasp
-npmDepsForWasp spec waspLibs =
-  N.NpmDepsForWasp
-    { N.waspDependencies =
-        Npm.Dependency.fromList
-          [ ("cookie-parser", "~1.4.6"),
-            ("cors", "^2.8.5"),
-            ("express", expressVersionStr),
-            ("morgan", "~1.10.0"),
-            ("dotenv", "^16.0.2"),
-            ("helmet", "^6.0.0"),
-            ("superjson", show superjsonVersion)
-          ]
-          ++ depsRequiredByWebSockets spec
-          ++ waspLibsNpmDeps,
-      N.waspDevDependencies =
-        Npm.Dependency.fromList
-          [ ("nodemon", "^2.0.19"),
-            -- TODO: Allow users to choose whether they want to use TypeScript
-            -- in their projects and install these dependencies accordingly.
-            ("typescript", show typescriptVersion),
-            ("@types/express", show expressTypesVersion),
-            ("@types/express-serve-static-core", show expressTypesVersion),
-            ("@types/node", "^" <> majorNodeVersionStr <> ".0.0"),
-            ("@tsconfig/node" <> majorNodeVersionStr, "latest"),
-            ("@types/cors", "^2.8.5"),
-            ("rollup", "^4.9.6"),
-            ("rollup-plugin-esbuild", "^6.1.1"),
-            ("@rollup/plugin-node-resolve", "^16.0.0")
-          ]
-    }
+npmDepsFromWasp :: AppSpec -> [WaspLib.WaspLib] -> N.NpmDepsFromWasp
+npmDepsFromWasp spec waspLibs =
+  N.NpmDepsFromWasp $
+    N.NpmDepsForPackage
+      { N.dependencies =
+          Npm.Dependency.fromList
+            [ ("cookie-parser", "~1.4.6"),
+              ("cors", "^2.8.5"),
+              ("express", expressVersionStr),
+              ("morgan", "~1.10.0"),
+              ("dotenv", "^16.0.2"),
+              ("helmet", "^6.0.0"),
+              ("superjson", show superjsonVersion)
+            ]
+            ++ depsRequiredByWebSockets spec
+            ++ waspLibsNpmDeps,
+        N.devDependencies =
+          Npm.Dependency.fromList
+            [ ("nodemon", "^2.0.19"),
+              -- TODO: Allow users to choose whether they want to use TypeScript
+              -- in their projects and install these dependencies accordingly.
+              ("typescript", show typescriptVersion),
+              ("@types/express", show expressTypesVersion),
+              ("@types/express-serve-static-core", show expressTypesVersion),
+              ("@types/node", "^" <> majorNodeVersionStr <> ".0.0"),
+              ("@tsconfig/node" <> majorNodeVersionStr, "latest"),
+              ("@types/cors", "^2.8.5"),
+              ("rollup", "^4.9.6"),
+              ("rollup-plugin-esbuild", "^6.1.1"),
+              ("@rollup/plugin-node-resolve", "^16.0.0")
+            ],
+        peerDependencies = []
+      }
   where
     majorNodeVersionStr = show (SV.major $ getLowestNodeVersionUserAllows spec)
 
     waspLibsNpmDeps = map (WaspLib.makeLocalNpmDepFromWaspLib libsRootDirFromServerDir) waspLibs
 
-genNpmrc :: Generator FileDraft
-genNpmrc =
-  return $
-    C.mkTmplFdWithDstAndData
-      (C.asTmplFile [relfile|npmrc|])
-      (C.asServerFile [relfile|.npmrc|])
-      Nothing
+genNpmrc :: AppSpec -> Generator [FileDraft]
+genNpmrc spec
+  -- We only use `.npmrc` to force `npm` to error out if the Node.js version is incompatible.
+  --
+  -- In dev mode, we already check the Node.js version ourselves before running any `npm` commands,
+  -- so we don't need this there.
+  --
+  -- We do expect users to manually go into the generated directories when bundling the built ouput.
+  -- So we do add the `.npmrc` there to help them avoid using an incompatible Node.js version.
+  | AS.isBuild spec =
+      return
+        [ C.mkTmplFdWithDstAndData
+            (C.asTmplFile [relfile|npmrc|])
+            (C.asServerFile [relfile|.npmrc|])
+            Nothing
+        ]
+  | otherwise =
+      return []
 
 genGitignore :: Generator FileDraft
 genGitignore =